move client init to new module
