+++ src/extension.ts
@@ -80,10 +80,7 @@ import { SchemasViewProvider } from "./viewProviders/schemas";
 import { SEARCH_DECORATION_PROVIDER } from "./viewProviders/search";
 import { SupportViewProvider } from "./viewProviders/support";
 import { TopicViewProvider } from "./viewProviders/topics";
-import { CCLOUD_CONNECTION_ID } from "./constants";
-import { LanguageClient, LanguageClientOptions, Trace } from "vscode-languageclient/node";
-import { WebSocket } from "ws";
-import { WebsocketTransport } from "./sidecar/websocketTransport";
+import { initializeLanguageClient } from "./flinkSql/languageClient";
 const logger = new Logger("extension");
 
 // This method is called when your extension is activated based on the activation events
@@ -135,7 +132,7 @@ export async function activate(
   }
 }
 
-let languageClient: LanguageClient;
+let languageClient;
 /**
  * Activate the extension by setting up all the necessary components and registering commands.
  * @remarks This is the try/catch wrapped function called from the extension's .activate() method
@@ -175,37 +172,16 @@ async function _activateExtension(
   // set up the preferences listener to keep the sidecar in sync with the user/workspace settings
   const settingsListener: vscode.Disposable = await setupPreferences();
   context.subscriptions.push(settingsListener);
-  const addr = `ws://127.0.0.1:26636/flsp?connectionId=${CCLOUD_CONNECTION_ID}&region=us-east1&provider=gcp&environmentId=env-x7727g&organizationId=f551c50b-0397-4f31-802d-d5371a49d3bf`;
-  const conn = new WebSocket(addr);
-  conn.onopen = async () => {
-    logger.info("WebSocket connection opened");
+
+  if (vscode.workspace.getConfiguration().get(ENABLE_FLINK, false)) {
     try {
-      const transport = new WebsocketTransport(conn);
-      let serverOptions = () => {
-        return Promise.resolve(transport);
-      };
-      const clientOptions: LanguageClientOptions = {
-        documentSelector: [
-          { scheme: "file", language: "plaintext" },
-          { scheme: "file", language: "flinksql" },
-        ],
-        outputChannelName: "FlinkSQL Language Server"
-      };
-      languageClient = new LanguageClient(
-        "flinksqlLanguageServer",
-        "FlinkSQL Language Server",
-        serverOptions,
-        clientOptions,
-      );
-      languageClient.setTrace(Trace.Verbose);
-      languageClient.start().then(() => {
-        logger.info("Language server started");
-      });
+      languageClient = await initializeLanguageClient();
+      logger.info("Language server started");
       context.subscriptions.push(languageClient);
     } catch (e) {
-      logger.error(`Error starting language server: ${e}`);
+      logger.error(`Error initializing FlinkSQL language client: ${e}`);
     }
-  };
+  }
 
   // set up the different view providers
   const resourceViewProvider = ResourceViewProvider.getInstance();
